tRestructure project layout and add test framework
We need a way to run command-line functional tests. This adds a basic
framework based on the golden files pattern used by the dep tool. In
order for this to work, we also move command line code into
a subdirectory and pull the dictionary portion out to a top-level
xkcdpwd package. Drop cobra in favor of go flags command-line ala the
dep project, because cobra's model doesn't let us test the command-line
both in-line and as a subprocess in parallel. We want to do this so we
can get coverage metrics and functionally test the binary.

The integration tests use a testdata directory in cmd/xkcdcmd. The test
harness recursively walks that directory looking for testscase.json
files. The test name is the path from testdata to the testcase.json
file.

The json file describes how to run the tests, and if the last test
command is expected to generate an error. Optionally, a file stdout.txt
records the expected output of the commmand.

Given that xkcdpwd generates random passphrases, in order for static
files to properly express the expected output, we use a test build tag
to bake in an english dictionary of only one word.